Swap out the <code> tag for <source>,  Add pygments
This allows the <code> tag to be used separately by using <source> for
code blocks intended to be formatted.

Additionally, if pygments is available, and use_pygments=True is passed
to html.make_parser, pygments is used to generate with
source-highlighting. The lang attribute will be used to lookup a lexer
for the language.